v0.3 (Arbitrage)
Updated and got working both the calculator and the sportsbook.

The Calculator is a dynamic arbitrage calculator that will tell you the profit payout between odds and set stakes of any opposing fights.

The Sportsbooks will pool the all bookmakers and dynamically calculate the best arbitrage for that matchup and display it.

If arbitrage exist it will show up.